+++ lib/workling/remote/runners/starling_runner.rb
@@ -17,11 +17,10 @@ module Workling
     module Runners
       class StarlingRunner < Workling::Remote::Runners::Base
         cattr_accessor :routing
-        cattr_accessor :client
         
         def initialize
-          StarlingRunner.client = Workling::Starling::Client.new
-          StarlingRunner.routing = Workling::Starling::Routing::ClassAndMethodRouting.new
+          StarlingRunner.client = Workling::Clients::MemcacheQueue.new
+          StarlingRunner.routing = Workling::Routing::ClassAndMethodRouting.new
         end
         
         # enqueues the job onto Starling.